20230723 백준 Study

+++ BOJ_17070.java
@@ -4,7 +4,7 @@ import java.io.InputStreamReader;
 import java.util.StringTokenizer;
 
 public class BOJ_17070 {
-    static int cnt = 0;
+    static long cnt = 0;
     static int n;
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
